ingestion-service/requirements.txt
fastapi
uvicorn
pika
python-multipart
boto3
python-dotenv
httpx
